Abstraction for providing pixels to the encoder.
Now there are two ways to feed the video encoder with pixeldata. Either
provide a method that fills the yuv buffers needed for encoding or
provide a slice to BGRA encoded data, 8 bits for each color. Like this
sws_scale can be used for BGRA which is faster than the previous method.