Do not let a queue of requests for frames build up.
Previously every GetFrame messages was guaranteed to get a video frame
in response as long as the connection is open. If the rate at which
frames were requested was too high a queue of requests built up and
added some lag to other requests like `VideoCommands::Start`.

This problem is solved by dropping frame requests if they start to queue
up.