Fix crashed on Radeon graphics with VAAPI
For some reason the radeon driver reports some pixel formats as
supported even though they are not. This lead to a crash in
avcodec_receive_packet and further down the stack vaEndPicture which at
first seemed unrelated to the pixel format.
For now the solution is to check if the vendor string reported by
vaQueryVendorString contains "Radeon" and if so, to just force NV12 as
pixel format.