Max Video Resolution now does what it says.
Previously some calculations have been made in order to allow stretching
the video while keeping full resolution for both width and height.
Unfortunately this sometimes results in videos with extreme width or
height (multi monitor setups), that don't play in some browsers. This is
why Max Video Resolution is now a hard limit on the video that Weylus
sends to the browsers. To preserve the ability to enjoy videos at full
resolution even if stretching, the Max Video Resolution can now be
greater than the available window resolution.