Prevent video from getting stuck.
Some browsers do not set the readyState of the video to 3 or above after
the window/tab is brought back from background. This prevented seeking
to the end of the video. This is now worked around by forcefully seeking
to the end of the video once there is a lag of more than 5 seconds.