Add selection of custom input areas on Wayland.
Automatic and correct input mapping on Wayland is difficult/impossible.
Thus, on Wayland Weylus now supports the selection of a custom input
area. This is done via a pop-up window that is positioned according to
the desired area of input. Typically, the different input devices
(mouse, touch, pen) can be mapped to the whole workspace (all screens)
or only a specific screen. KDE for example offers input mapping of touch
input to a specific screen. To Weylus this specific mapping is not know
and needs to be specified in a choice menu of the pop-up window.