Workaround for hovering pens and touch rejection
Some browsers do not send a pointer event to indicate that a hovering
pen has left the hovering range. This results in the pen still being
active, even if there are touch events and the pen is lifted. In this
case we manually remove the pen to prevent touch rejection.